Fix typos, fix logging_time function
Fix some typos that were causing errors and incorrect formatting in txt file.
Change logging_time function so that it returns a string for the caller to write to file, instead of writing directly to file.
Change file mode to write instead of append.

To do: File is being saved in the wrong folder. Some minor formatting mistakes in txt file. Support for multiple acquisitions needs to be added.