Move sampling operation to seperate thread
Long sampling operations used to cause the program to be marked as 'Not responding'. This is fixed by moving the sampling operation to a seperate thread.
Window.refresh() is called periodically when printing data, since printing takes a long time and might also cause the app to become unresponsive.

A problem that still persists is that taking a large number of samples with many channels causes the sampling function to block. This might be a problem with STM32-side program. Will have to look into that further.